backend/monitoring/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.1
Django==3.2.12
django-seed==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
Faker==12.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toposort==1.7
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
